Remove original test location
